requirements.txt
selenium==4.17.2
beautifulsoup4==4.12.3
schedule==1.2.1
python-telegram-bot==20.7